ManagerHCI odbrana

+++ Hospital/Hospital/Service/ItemInRoomService.cs
@@ -63,10 +63,10 @@ namespace Hospital.Service
         public ItemInRoom MoveWholeItemNowToMainStorage(ItemInRoom itemInRoom)
         {
             Room mainStorage = roomRepository.GetRoomById(5);
-            itemInRoom.room_id = mainStorage.Id;
+            itemInRoom.room_id = (int)mainStorage.Id;
             itemInRoom.room.Id = mainStorage.Id;
 
-            ItemInRoom itemInDestinationRoom = AlreadyExists(itemInRoom, mainStorage.Id);
+            ItemInRoom itemInDestinationRoom = AlreadyExists(itemInRoom, (int)mainStorage.Id);
             if (itemInDestinationRoom != null)
             {
                 itemInDestinationRoom.Quantity += itemInRoom.Quantity;
@@ -94,7 +94,7 @@ namespace Hospital.Service
 
         private bool Move(ItemInRoom itemInRoom, Room destinationRoom, uint quantity)
         {
-            ItemInRoom itemInDestinationRoom = AlreadyExists(itemInRoom, destinationRoom.Id);
+            ItemInRoom itemInDestinationRoom = AlreadyExists(itemInRoom, (int)destinationRoom.Id);
             if (itemInDestinationRoom != null)
             {
                 itemInDestinationRoom.Quantity += quantity;
